FULL PIPELINE START TO END_PO/Dashboard Summary and Reporting/fortifai-reports/requirements.txt
fastapi
uvicorn[standard]
pandas
python-dateutil
openpyxl
xlsxwriter
psycopg2-binary
azure-storage-blob
